-n-t-p-main/requirements.txt
flask
flask-cors
google-genai
python-dotenv

